Create post.html
Creates a new layout for blog posts called post.html and a folder to store each individual post called _posts/. Creates the layout. Creates a file named post.html in the _layouts folder. Notice the post layout uses the default layout as it's base, and adds a couple new liquid tags to print the title of the post and date.