Create 2018-03-08-akeba-mitchell-site-launched.md
Make a _posts/ directory where blog posts will be stored. Inside that folder will be the first post. Jekyll is very strict with how these files are named, so pay attention. It must follow the convention YYYY-MM-DD-title-of-my-post.md. This file name gets translated into the permalink for the blog post. Creates a file named 2018-03-08-akeba-mitchell-site-launched.md. Note the file extension .md stands for Markdown, and the Markdown syntax used inside the file gets converted to HTML by Jekyll. Like Wikitext, Markdown is a markup language with a syntax closer to plain text. The idea of Markdown is to get out of the author's way so they can write their HTML content quickly, making Markdown very suitable as a blog authoring syntax.